tools/curseforge-instance-parser/pyproject.toml
[project]
name = "curseforge-instance-parser"
version = "0.1.0"
description = "Parse CurseForge instance file(minecraftinstance.json) and detect added/updated mods."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pyyaml>=6.0.3"]
